Check if file exists, otherwise it destroys it.

+++ Projet.Net/Program.cs
@@ -19,10 +19,22 @@ namespace Projet.Net
 
             // Gives the path of the Documents folder on the computer
             String path = System.Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments) + "\\" + appName;
+            
             // I guess we could create the folder if it's not there on loading the appWindow ?
             Directory.CreateDirectory(path);
             // Then create the non existing .json
-            File.Create(path + "\\workspace.json" );
+            if(!File.Exists(path + "\\workspace.json"))
+                File.Create(path + "\\workspace.json" );
+
+            /* We should put at least in it
+             * "tags" : [
+             * "Vacances",
+             * "Mer",
+             * "Montagnes"
+             * ],
+             * "images" : [
+             * ]
+             */
 
             Base.getInstance().loadWorkspace();
             Application.EnableVisualStyles();
